+++ Dagger2Kotlin/src/main/java/com/example/android/dagger/di/AppComponent.kt
@@ -1,8 +1,9 @@
 package com.example.android.dagger.di
 
 import android.content.Context
+import com.example.android.dagger.login.LoginComponent
 import com.example.android.dagger.main.MainActivity
-import com.example.android.dagger.settings.SettingsActivity
+import com.example.android.dagger.registration.RegistrationComponent
 import dagger.BindsInstance
 import dagger.Component
 import javax.inject.Singleton
@@ -45,11 +46,17 @@ interface AppComponent {
      * @return RegistrationComponent.Factory
      */
     // Classes that can be injected by this Component
-   // Expose RegistrationComponent factory from the graph
+
     fun registrationComponent(): RegistrationComponent.Factory
 
+    /**
+     *  Expose loginComponent factory from the graph
+     * @return LoginComponent.Factory
+     */
+    fun loginComponent(): LoginComponent.Factory
+
+
     fun inject(activity: MainActivity)
-    fun inject(activity: SettingsActivity)
 
 
 }
