+++ Dagger2Kotlin/src/main/java/com/example/android/dagger/di/AppSubcomponents.kt
@@ -1,8 +1,10 @@
 package com.example.android.dagger.di
 
+import com.example.android.dagger.login.LoginComponent
+import com.example.android.dagger.registration.RegistrationComponent
 import dagger.Module
 
 // This module tells AppComponent which are its subcomponents
-@Module(subcomponents = [RegistrationComponent::class])
+@Module(subcomponents = [RegistrationComponent::class, LoginComponent::class])
 class AppSubcomponents {
 }
