+++ Dagger2Kotlin/src/main/java/com/example/android/dagger/login/LoginActivity.kt
@@ -29,18 +29,24 @@ import com.example.android.dagger.main.MainActivity
 import com.example.android.dagger.MyApplication
 import com.example.android.dagger.R
 import com.example.android.dagger.registration.RegistrationActivity
+import javax.inject.Inject
 
 class LoginActivity : AppCompatActivity() {
 
-    private lateinit var loginViewModel: LoginViewModel
+    // 1) LoginViewModel is provided by Dagger
+    @Inject lateinit var loginViewModel: LoginViewModel
     private lateinit var errorTextView: TextView
 
     override fun onCreate(savedInstanceState: Bundle?) {
+        // 2) Creates an instance of Login component by grabbing the factory from the app graph
+        // and injects this activity to that Component
+        (application as MyApplication).appComponent.loginComponent().create().inject(this)
+
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_login)
 
-        // Creates ViewModel and listens for the loginState LiveData
-        loginViewModel = LoginViewModel((application as MyApplication).userManager)
+        // // Creates ViewModel and listens for the loginState LiveData
+        // loginViewModel = LoginViewModel((application as MyApplication).userManager)
         loginViewModel.loginState.observe(this, { state ->
             when (state) {
                 is LoginSuccess -> {
