+++ Dagger2Kotlin/src/main/java/com/example/android/dagger/login/LoginComponent.kt
@@ -1,0 +1,21 @@
+package com.example.android.dagger.login
+
+import com.example.android.dagger.di.ActivityScope
+import dagger.Subcomponent
+
+// Scope annotation that the LoginComponent uses
+// Classes annotated with @ActivityScope will have a unique instance in this Component
+@ActivityScope
+// Definition of a Dagger subcomponent
+@Subcomponent
+interface LoginComponent {
+
+    // Factory to create instances of LoginComponent
+    @Subcomponent.Factory
+    interface Factory {
+        fun create(): LoginComponent
+    }
+
+    // Classes that can be injected by this Component
+    fun inject(activity: LoginActivity)
+}
