+++ Dagger2Kotlin/src/main/java/com/example/android/dagger/login/LoginViewModel.kt
@@ -19,12 +19,13 @@ package com.example.android.dagger.login
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import com.example.android.dagger.user.UserManager
+import javax.inject.Inject
 
 /**
  * LoginViewModel is the ViewModel that [LoginActivity] uses to
  * obtain information of what to show on the screen and handle complex logic.
  */
-class LoginViewModel(private val userManager: UserManager) {
+class LoginViewModel @Inject constructor(private val userManager: UserManager) {
 
     private val _loginState = MutableLiveData<LoginViewState>()
     val loginState: LiveData<LoginViewState>
