+++ Dagger2Kotlin/src/main/java/com/example/android/dagger/registration/RegistrationComponent.kt
@@ -1,6 +1,6 @@
-package com.example.android.dagger.di
+package com.example.android.dagger.registration
 
-import com.example.android.dagger.registration.RegistrationActivity
+import com.example.android.dagger.di.ActivityScope
 import com.example.android.dagger.registration.enterdetails.EnterDetailsFragment
 import com.example.android.dagger.registration.termsandconditions.TermsAndConditionsFragment
 import dagger.Subcomponent
